web/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.7
openai==1.12.0
boto3==1.34.22
pytest==7.4.4
gunicorn==21.2.0
aiofiles
email-validator

# Development dependencies
flake8==7.1.2
black==25.1.0
isort==6.0.1
mypy==1.15.0
